requirements.txt
streamlit
fastapi
uvicorn
requests
python-dotenv
sentence-transformers
pydantic-settings
pinecone-client
scikit-learn
pandas
matplotlib
pinecone
streamlit-option-menu
prophet
google-generativeai
python-dotenv
PyPDF2
python-docx